+++ tes-rest-models/src/main/java/dk/sunepoulsen/tes/rest/models/HashMapModel.java
@@ -1,0 +1,12 @@
+package dk.sunepoulsen.tes.rest.models;
+
+import java.util.HashMap;
+
+/**
+ * Model for dynamic json.
+ * <p>
+ *     It can represent any type of json.
+ * </p>
+ */
+public class HashMapModel extends HashMap<String, Object> implements BaseModel {
+}